Undo splitting of test files
Since we now run tests fully in parallel this only creates more
confusion. We should instead clearly tag tests to make it clear who
tests what, see #313.
